chore: Revert "feat: support 64-bit signed integers"
This reverts commit 3990fc5edfce76f4c26adaae652187691a84156e.

+++ src/main/java/me/mtk/torrey/frontend/ast/ConstantConvertable.java
@@ -7,5 +7,5 @@ package me.mtk.torrey.frontend.ast;
  */
 public interface ConstantConvertable
 {
-  public long toConstant();
+  public int toConstant();
 }
